Add initial framing for a contributing doc
We have contributor documentation, that explains placement
from within, but we don't have much in the way of docs that
explain how to make contributions and the guidelines for
how do it well. This patch is the first in a series that will
add that information.

This first patch simply creates an initial contributing page
with a light intro and titles for four initial sections.

Change-Id: Iea127a052a42ed5cc5349b89137381f782717af3