Fill in reviewing section of contributing.rst
This provides some links to more info about reviewing, and some
guidelines on how to be a good reviewer, with special
considerations for cores.

Not included is information on how a patch submitter should hassle
someone to get attention. With the low volume of patches that go
through placement, that should not be something the submitter should
have to care about. It's the job of regular reviewers to be aware
of new stuff. The guideline about latency covers that.

Change-Id: I61cf791c956a75304b87ffb4a16a8252fc36800b