Run nova-tox-functional-py36 in the placement gate
This requires some slight adjustsments to the job in nova,
found in the depends-on.

The result of this will be _this_ placement change running
against nova master.

In order to blacklist the api_sample_tests (to save some time)
the job is inherited and some args are added to the tox command.

Story: 2005562
Task: 30746
Depends-On: https://review.opendev.org/657074
Change-Id: I72ae7f552879be13777fea90fa7055e056d309a7