Skip notification sample tests when running nova functional
The notification sample tests are many and at least some of
them contribute quite a bit of time so may as well not run
them when placement runs the nova functional tests.

In a recent run those tests took ~11.5 minutes to run in total,
with ~6.5 minutes being devoted to the actual tests.

Tuning this to perfection is probably a fools errand but
making tweaks when there are obviously useless (in this
context) time consumers, may as well kill them.

(Not listing a task here as the story has already merged
so the following syntax will operate similar to related-bug.)

Change-Id: I8f762cbf0ce2126609b7c24a8054785e252bdd03
Story: 2005562