Remove null root provider protections
A when root providers were added, a series of TODOs were
left in allocation candidate and resource provider data
handling. These TODOs should be removed once we are assured
that there are no null roots. That is the case now, given
that we had an online migration and a status check and now
(prior to this patch) there is a blocker migration and the
status check is an error instead of warning.

The release note for the change of the status check should
suffice for covering this. That is: it makes what it says,
True.

Story: 2005613
Task: 30860
Change-Id: I84baff29505550f3f20069ad5817784d0d1aaea6