Raise os-traits os-resource-classes constraints
Because os-traits and os-resource-classes are libraries that
provide enumerations, we usually want whatever the latest
version is, even in lower-constraints.txt. We haven't, however,
been good about keeping those up to date.

This updates both requriements.txt and lower-constraints.txt to
the latest versions of both os-traits and os-resource-clases.

No change is required is global upper-constraints. That already
has the latest.

Change-Id: I0bd0e8e071d6275c3e21556018d476c97e8533ae
Story: 2005651
Task: 30939