Add RequestGroupSearchContext class
There are cases where exactly the same database query is issued several
times in a request group, such as provider_ids_matching_aggregates().

This patch creates a new object, `RequestGroupSearchContext` to cache
the results not to ask to issue the same query in a request group.

The main refatctoring to reduce the number of the same query will be
done in following patches.

This patch brings a slight optimzation to skip retrieving sharing
providers for `use_same_provder=True` case.

Change-Id: Ifbaac9861f86d85a5bff58573c30a4cf957503d8
Story: 2005712
Task: 31038