Enhance debug logging in allocation candidate handling
There is reasonably good debug logging when filtering results to
create a set of allocation candidates, but the latter half of the
processing which merges candidates, removes nested providers and
limits results is not well logged.

This change adds some debug logs in that handling to provide more
information. While the logs do report the sizes of the entities
involved another important factor of the logging is the timestamps
that indicate time spent between each message.

Change-Id: I98ce4cade9acd64285c5a65bd439e37cb6a308f3
Story: 2005647
Task: 30927