Avoid traversing summaries in _check_traits_for_alloc_request
While doing other work, I noticed that _check_traits_for_alloc_request
traverses a list of summaries multiple time to find the "right one".
This is a classic case of a list being the wrong data structure for
the job.

Instead use the request resource resource provider's id to key into
the summaries dict directly to get the relevant summary.

Then, since we know the summary has the trait info we need, we don't
need to pass prov_traits to the method.

Change-Id: Ic2f8f7bdc2984db2011cec329fb6f5b9efec5a0f