Use trait strings in ProviderSummary objects
When ProviderSummary objects are created and then later serialized
to JSON the traits attribute has been a list of Trait objects.

At no point in the processing of that attribute does any caller want
it to be a Trait. They are stored as TraitS but used as strings.

Therefore, this change makes it so it is a list of strings, avoiding
a few different instances of translating one way or the other, keeping
the data in the format the system wants.

Change-Id: Ia9d81ce87111ec3496d10ed254f069c04b9bdf3c