Canary test for os-traits version
Since [1] we have decided to pin placement to the latest versions of
os-resource-classes and os-traits. Turns out we had gabbi tests that
would fail as soon as the os-resource-classes upper constraint was
bumped, alerting us that it was time to sync that requirement. This
commit adds a similar test for os-traits.

[1] https://review.opendev.org/#/c/658419/

Change-Id: I57e834311c73282a2e7a00b14823812cb6f3e1e6