Don't run functional.db tests in nova functional run
To further refine the tests being run during the nova functional
test job, remove the functional.db tests. These shouldn't be
interacting with placement and because some of them use live
mysql or postgresql services for the migration tests, are more
susceptible to intermittent failures when nodes are being slow.

Change-Id: I2cfc6cb74d3acf78043a20cc7254d1315d6fabf0