Uniquify allocation mappings
When creating the allocation mappings output, use a set() not a
list() so that any single resource provider only shows up once.
The initial implementation (at
Ie78ed7e050416d4ccb62697ba608131038bb4303) allowed a provider to
show up multiple times if it contributed more than one class
of inventory.

This slipped through because of a missing test. A test that covers
the issue is added.k

Change-Id: If00f01534b7d0ec84ca8abaeef5b90a16cbcffc3