RequestWideParams and RequestWideSearchContext
Recent design discussions have reified the concept that a request to GET
/allocation_candidates has two kinds of query parameters:

- Those that apply to request groups: resources[$S], required[$S], etc.
  Essentially all the ones that can be (optionally) suffixed.
- Those that apply to the request as a whole: limit, group_policy.

Group-specific parameters are already encapsulated in the
placement.lib.RequestGroup class, one instance of which is used to
represent each request group in the query. And then database and other
filtering logic is encapsulated in
placement.objects.research_context.RequestGroupSearchContext.

This commit pulls the existing request-wide parameters, limit and
group_policy, into a new class, placement.lib.RequestWideParams; and
introduces placement.objects.research_context.RequestWideSearchContext
to encompass the filtering logic for the same.

In so doing, we adjust the signatures of methods which used to use those
individual parameters.

At the same time, parameters and local variables referring to the list
of RequestGroup are renamed from `requests` to `groups` to
disambiguate/clarify what they are.

In the end, this is just a refactor. No logic is changed.

Change-Id: I572ec8e6c0f9d81ead49964f41b46ab68a3de9a6