Remove overly-verbose allocation request log
In I952d5229d6c40588cde6197683117a7e19127939 an "got %d
allocation requests under root provider %s" log message was added.
It reports a log line for every allocation request under every
root provider in the candidates. In a large but sparsely used
cloud this generates a large number of log lines (see the identified
story for more details and links to examples), well more than 75%
of the lines in a log.

The information added by this particular line is not nearly as
useful as the other lines added in the change identified above.
So here, in this change, it is removed.

Change-Id: I4aee9de5b19045e6e773349e8f77ccb7f359d360
Story: 2005918
Task: 34179