Recreate: incorrect mappings with group_policy=none
Adds a gabbi test demonstrating the bug in the associated story:

GET /allocation_candidates
with group_policy=none
and two request groups
which consolidate to land on the same provider

In this case, the mappings should show two entries: one from each
request group to the (same) provider.

However, the test demonstrates that only one mapping is returned. (I
think it's arbitrary which request group will be represented, so the
test only asserts that the size of the mappings dict is one instead of
two.)

Story: #2006068
Task: #34786

Change-Id: I0b12a8ac80e42c605cfd26a3adc1312dedaa3f99