Fix up some inaccuracies in perfload comments and logs
The review of the addition of nested perfload (in
I617161fde5b844d7f52dc766f85c1b9f1b139e4a ) identified some
inaccuracies in the comments and logs. This fixes some of
those.

It does not, however, fix some of the duplication between the
two runner scripts. This will be done later.

Change-Id: I9c57125e818cc583a977c8155fcefcac2e3b59df