Add consumer_types migration, database and object changes
In advance of a microversion 1.38, add the necessary database and
object code to support saving consumer types when creating consumers
and reporting those types when requesting usage information.

This patch adds a new table called ``consumer_types`` and also
add a new column called ``consumer_type_id`` that references the id
in the consumer_types table.

The consumer_type_id is nullable and records with a NULL
consumer_type_id are considered as 'unknown'.

In future changes we may wish to AttributeCache consumer types in
the same way that traits and resource classes are cached.

Usage object can now be retrieved by consumer type and include
consumer type information when retrieved. Until 1.38 that info
will not be used in results.

Allocations are written and retrieved with a consumer type but
the data is not yet presented in HTTP requests.

Co-Authored-By: Chris Dent <cdent@anticdent.org>
Co-Authored-By: melanie witt <melwittt@gmail.com>

Story: 2005473
Task: 35683
Task: 36420

Change-Id: Ic74c3d1b8aa272e9c7d1070fe7658b928e693889