Simplify placement.microversion:_fully_qualified_name
_fully_qualified_name was copied from Twisted. The original
version does more than Placement needs. We only need to
handle module-level classes and functions (and really only
functions). No nested of those, nor object methods in any
form.

This is because we only use the method for finding the
right microversioned handler of a method with the same
name and placement HTTP handlers are universally functions.

The previous version of the method had code that was never
called. Here, that dead code is removed and the remainder
is simplified.

If it is called incorrectly there will be an (intentionally)
untrapped TypeError during running of the tests.

N.B.: This was discovered while reviewing coverage data.

Change-Id: Id8f96801116818dda6c68c301619aee185e8fbff