Use integrated-gate-placement zuul template
The integrated-gate-placement adds the tempest-integrated-placement
and grenade-py3 jobs. tempest-integrated-placement runs a subset of
tempest tests which are relevant for placement, e.g. it does not
run keystone tests.

The integrated-gate-py3 template is replaced by this change. It
specified the tempest-full-py3, meaning many tests not relevant
to placement were run.

See the depends-on for more information on the rationale.

Depends-On: https://review.opendev.org/669309
Change-Id: I509ded97085fbae5a0cc55fe66f6e30e638b72bc