Optimize trait creation to check existence first
Comparing benchmarks of creating (repeated) resource classes with
traits, it became clear that creating resource classes [1] was about
1/3rd faster. Comparing the code, they were very similar, however
the resource class side was ordered to check for existence first
before attempting to do a create(). This meant that in the fairly common
case where a resource class already existed we could finish early.

This change updates the code so that the common idiom used in clients
of "ensuring a trait" with PUT /traits/$name routine to be the same
as the "ensuring a resource class" with PUT /resource_classes/$name
routine: check that it is already there, first. Since it is best
practice to always ensure trait or resource class before using a
CUSTOM_ one, this is a good orientation.

Note that with the recent addition of the AttributeCache, these queries
for get_by_name [2] go to the ResourceClass or Trait cache. However, at
this stage in processing, the cache is empty and calls to get_by_name
will fill it.

A unit test is added to confirm that the 'put_trait' handler will look
at get_by_name and then go back to assuming the trait exists if the
create() fails because it already exists. The general behavior of
whether a trait is created or updated is confirmed by the gabbit
tests in trait.yaml. That test also confirms that we have a legit
last-modified time when we fall through both exceptions and make
the assumption that a trait was created by some other thread. It is
safe to use 'now' as the last-modified time because it was created
in this same second, just not by us.

[1] https://opendev.org/openstack/placement/src/commit/2f56f379a568d427ac9057086386074a7801b00c/placement/handlers/resource_class.py#L209
[2] https://opendev.org/openstack/placement/src/commit/2f56f379a568d427ac9057086386074a7801b00c/placement/objects/resource_class.py#L72

Change-Id: I88e42ef18bd5b2616a93730241d2419c6b431276