Use TraitCache for Trait.get_by_name
Change Trait.get_by_name to access the TraitCache. This was
something that was intended to be done in the patch that added
the cache [1] but was left out.

For most uses of get_by_name this has little impact: the cache
will be cold when the method is called, warmed to satisfy the
request. However, this makes Trait.get_by_name consistent with
ResourceClass.get_by_name.

[1] Ia19ea2b4ecdde25323579edf60ad6269d05e75a2

Change-Id: Iaa20e5e92e793b75f7773f86407cdedfb6584dab