Remove ProviderIds namedtuple
It turns out that all the callers of methods that return ProviderIds
access the data using attributes that are already present on the object
provided by a sql alchemy RowProxy object (which is the result of
iterating a ResultProxy object, returned by sqla.execute()). Creating
a namedtuple is redundant and adds some processing cost [1].

However, the AnchorIds namedtuple used in the same file _is_ necessary
because it provides a hashable type required by the set that wraps
the collection of AnchorIds.

The other user of namedtuple, RPCandidate, in
placement/objects/rp_candidates.py has a specific interface for which
a namedtuple is correct.

[1] provider_ids_from_rp_ids was identified as being a major contributor
when doing performance profiling. Most of the cost is executing with the
database, but the tuple creation is a factor.

Change-Id: I1e1522eae7b4360cf57c34164612ae228fad14a3