Remove double join in provider_ids_from_rp_ids
provider_ids_from_rp_ids is only used in _build_provider_summaries.
It is passed a collection of usages dict created
_get_usages_by_provider_tree this has usage information for _all_
the providers in the contributing trees, even those providers that
do not provide any resources (otherwise we wouldn't be able to
build a complete summary).

This means that the list of rp ids passed to provider_ids_from_rp_ids
includes all the relevant rps and the returned data structure can
itself be used to find root_provider_uuid and parent_provider_uuid,
instead of doing the join.

This speeds up the database query. Combined with the expanding
bindparam in the previous patch, this cleans up provider_ids_from_rp_ids
nicely.

Change-Id: If937053a8af3f0eecdd90aa807be0abc316d8c4f